Arreglando la actualización de roles de un usuario (#5452)
Se arregla el bug que no permitía actualizar los roles de un usuario en específico.
Resulta que las APIs que agregaban o removían roles a los usuarios no estaban 
funcionando correctamente, debido a que nunca se validaba el usuario al que 
se le iba a aplicar el nuevo rol. 

Quiere decir que el cambio se lo estaba aplicando al usuario que estaba logueado, 
pero al recargar la página si obtenía los roles del usuario a consultar y en ahí se
mostraba todo en blanco debido a que, en efecto, ese usuario no tenía ningún 
rol asignado.

Lo mismo estaba pasando con los experimentos, aunque estos no estaban
mostrando el error de los roles.

Por cierto, el error que salía ya va a aparecer con un mensaje más claro en caso de
que llegara a ocurrir nuevamente.

![UserRoles](https://user-images.githubusercontent.com/3230352/120766380-ce21ba00-c4df-11eb-9fec-870dbdb688dd.gif)


Fixes: #5450

+++ frontend/server/src/Controllers/User.php
@@ -3092,19 +3092,40 @@ class User extends \OmegaUp\Controllers\Controller {
      * @return array{status: string}
      *
      * @omegaup-request-param string $role
+     * @omegaup-request-param string $username
      */
     public static function apiAddRole(\OmegaUp\Request $r): array {
         \OmegaUp\Controllers\Controller::ensureNotInLockdown();
         $r->ensureMainUserIdentity();
-        \OmegaUp\Validators::validateStringNonEmpty($r['role'], 'role');
 
-        $role = self::validateAddRemoveRole($r->identity, $r['role']);
+        $role = $r->ensureString('role');
+        $role = self::validateAddRemoveRole($r->identity, $role);
+        $username = $r->ensureString(
+            'username',
+            fn (string $username) => \OmegaUp\Validators::usernameOrEmail(
+                $username
+            )
+        );
+        $user = self::resolveTargetUser($r);
+        if (is_null($user) || is_null($user->user_id)) {
+            throw new \OmegaUp\Exceptions\NotFoundException('userNotExist');
+        }
 
-        \OmegaUp\DAO\UserRoles::create(new \OmegaUp\DAO\VO\UserRoles([
-            'user_id' => $r->user->user_id,
-            'role_id' => $role->role_id,
-            'acl_id' => \OmegaUp\Authorization::SYSTEM_ACL,
-        ]));
+        try {
+            \OmegaUp\DAO\UserRoles::create(new \OmegaUp\DAO\VO\UserRoles([
+                'user_id' => $user->user_id,
+                'role_id' => $role->role_id,
+                'acl_id' => \OmegaUp\Authorization::SYSTEM_ACL,
+            ]));
+        } catch (\Exception $e) {
+            if (\OmegaUp\DAO\DAO::isDuplicateEntryException($e)) {
+                throw new \OmegaUp\Exceptions\DuplicatedEntryInDatabaseException(
+                    'userAlreadyHasSelectedRole',
+                    $e
+                );
+            }
+            throw $e;
+        }
 
         return [
             'status' => 'ok',
@@ -3117,16 +3138,27 @@ class User extends \OmegaUp\Controllers\Controller {
      * @return array{status: string}
      *
      * @omegaup-request-param string $role
+     * @omegaup-request-param string $username
      */
     public static function apiRemoveRole(\OmegaUp\Request $r): array {
         \OmegaUp\Controllers\Controller::ensureNotInLockdown();
         $r->ensureMainUserIdentity();
-        \OmegaUp\Validators::validateStringNonEmpty($r['role'], 'role');
 
-        $role = self::validateAddRemoveRole($r->identity, $r['role']);
+        $role = $r->ensureString('role');
+        $role = self::validateAddRemoveRole($r->identity, $role);
+        $username = $r->ensureString(
+            'username',
+            fn (string $username) => \OmegaUp\Validators::usernameOrEmail(
+                $username
+            )
+        );
+        $user = self::resolveTargetUser($r);
+        if (is_null($user) || is_null($user->user_id)) {
+            throw new \OmegaUp\Exceptions\NotFoundException('userNotExist');
+        }
 
         \OmegaUp\DAO\UserRoles::delete(new \OmegaUp\DAO\VO\UserRoles([
-            'user_id' => $r->user->user_id,
+            'user_id' => $user->user_id,
             'role_id' => $role->role_id,
             'acl_id' => \OmegaUp\Authorization::SYSTEM_ACL,
         ]));
@@ -3193,6 +3225,7 @@ class User extends \OmegaUp\Controllers\Controller {
      * Adds the experiment to the user.
      *
      * @omegaup-request-param string $experiment
+     * @omegaup-request-param string $username
      *
      * @return array{status: string}
      */
@@ -3203,8 +3236,19 @@ class User extends \OmegaUp\Controllers\Controller {
             throw new \OmegaUp\Exceptions\ForbiddenAccessException();
         }
 
+        $username = $r->ensureString(
+            'username',
+            fn (string $username) => \OmegaUp\Validators::usernameOrEmail(
+                $username
+            )
+        );
+        $user = self::resolveTargetUser($r);
+        if (is_null($user) || is_null($user->user_id)) {
+            throw new \OmegaUp\Exceptions\NotFoundException('userNotExist');
+        }
+
         \OmegaUp\DAO\UsersExperiments::create(new \OmegaUp\DAO\VO\UsersExperiments([
-            'user_id' => $r->user->user_id,
+            'user_id' => $user->user_id,
             'experiment' => $r->ensureEnum(
                 'experiment',
                 \OmegaUp\Experiments::getInstance()->getAllKnownExperiments()
@@ -3218,6 +3262,7 @@ class User extends \OmegaUp\Controllers\Controller {
      * Removes the experiment from the user.
      *
      * @omegaup-request-param string $experiment
+     * @omegaup-request-param string $username
      *
      * @return array{status: string}
      */
@@ -3228,8 +3273,19 @@ class User extends \OmegaUp\Controllers\Controller {
             throw new \OmegaUp\Exceptions\ForbiddenAccessException();
         }
 
+        $username = $r->ensureString(
+            'username',
+            fn (string $username) => \OmegaUp\Validators::usernameOrEmail(
+                $username
+            )
+        );
+        $user = self::resolveTargetUser($r);
+        if (is_null($user) || is_null($user->user_id)) {
+            throw new \OmegaUp\Exceptions\NotFoundException('userNotExist');
+        }
+
         \OmegaUp\DAO\UsersExperiments::delete(
-            $r->user->user_id,
+            $user->user_id,
             $r->ensureEnum(
                 'experiment',
                 \OmegaUp\Experiments::getInstance()->getAllKnownExperiments()
@@ -3877,6 +3933,90 @@ class User extends \OmegaUp\Controllers\Controller {
             ],
         ];
         return $response;
+    }
+
+    /**
+     * @return array{entrypoint: string, smartyProperties: array{payload: UserDetailsPayload, title: \OmegaUp\TranslationString}}
+     *
+     * @omegaup-request-param string $username
+     */
+    public static function getUserDetailsForTypeScript(\OmegaUp\Request $r) {
+        $r->ensureMainUserIdentity();
+        $username = $r->ensureString(
+            'username',
+            fn (string $username) => \OmegaUp\Validators::usernameOrEmail(
+                $username
+            )
+        );
+
+        $user = self::resolveTargetUser($r);
+        if (is_null($user) || is_null($user->user_id)) {
+            throw new \OmegaUp\Exceptions\NotFoundException('userNotExist');
+        }
+
+        $emails = \OmegaUp\DAO\Emails::getByUserId($user->user_id);
+        $emailsList = [];
+        foreach ($emails as $email) {
+            if (is_null($email->email)) {
+                continue;
+            }
+            $emailsList[] = $email->email;
+        }
+
+        $userExperiments = \OmegaUp\DAO\UsersExperiments::getByUserId(
+            $user->user_id
+        );
+        $userExperimentsList = [];
+        foreach ($userExperiments as $userExperiment) {
+            if (is_null($userExperiment->experiment)) {
+                continue;
+            }
+            $userExperimentsList[] = $userExperiment->experiment;
+        }
+
+        // TODO: Also support GroupRoles.
+        $systemRoles = \OmegaUp\DAO\UserRoles::getSystemRoles($user->user_id);
+
+        $roles = \OmegaUp\DAO\Roles::getAll();
+        $rolesList = [];
+        foreach ($roles as $role) {
+            if (is_null($role->name)) {
+                continue;
+            }
+            $rolesList[] = ['name' => $role->name];
+        }
+
+        $systemExperiments = [];
+        /** @var array<string, mixed> */
+        $defines = get_defined_constants(true)['user'];
+        foreach (\OmegaUp\Experiments::getInstance()->getAllKnownExperiments() as $experiment) {
+            $systemExperiments[] = [
+                'name' => $experiment,
+                'hash' => \OmegaUp\Experiments::getExperimentHash($experiment),
+                'config' => \OmegaUp\Experiments::getInstance()->isEnabledByConfig(
+                    $experiment,
+                    $defines
+                ),
+            ];
+        }
+
+        return [
+            'smartyProperties' => [
+                'payload' => [
+                    'emails' => $emailsList,
+                    'experiments' => $userExperimentsList,
+                    'roleNames' => $rolesList,
+                    'systemExperiments' => $systemExperiments,
+                    'systemRoles' => $systemRoles,
+                    'username' => $username,
+                    'verified' => $user->verified != 0,
+                ],
+                'title' => new \OmegaUp\TranslationString(
+                    'omegaupTitleAdminUsers'
+                ),
+            ],
+            'entrypoint' => 'admin_user',
+        ];
     }
 
     /**
